Update README.md
Caroline and Ria's final computer science project. This project replicates the searching and filtering system on Netflix. 
The data.py file defines out class Movie, which contains different objects relevant to a movie. 
The functions.py file is our driver file. It contains all of the functions we will test our data with. 
The functions.test.py file contains our unit tests for the functions. 
The user.input.py file contains a function that directoy takes user input and returns a result. 
The titles.txt file is a text file which we reference in our function.